First pass at re-modularizing non-essential functions
