+++ app.js
@@ -1,0 +1,5 @@
+﻿var Nightmare = require("./lib/nightmare");
+require("./actions/core");
+require("./actions/emulateInput");
+require("./actions/cookies");
+module.exports = Nightmare;
