--- lib/app.js
@@ -1,3 +1,0 @@
-﻿var Nightmare = require("./nightmare");
-require("./cookies");
-module.exports = Nightmare;
