+++ lib/nightmare.js
@@ -161,14 +161,6 @@ class Nightmare {
     }
 
     /**
-     * Go back to previous url.
-     */
-    back() {
-        debug('.back()');
-        return this._invokeRunnerOperation("goBack");
-    };
-
-    /**
      * Creates a nightmare object which can be used to chain a number of actions sequentally.
      */
     chain() {
@@ -274,141 +266,6 @@ class Nightmare {
     }
 
     /**
-     * Check a checkbox, fire change event
-     *
-     * @param {String} selector
-     */
-    check(selector) {
-        debug('.check() ' + selector);
-        return this.evaluate_now(function (selector) {
-            var element = document.querySelector(selector);
-            var event = document.createEvent('HTMLEvents');
-            element.checked = true;
-            event.initEvent('change', true, true);
-            element.dispatchEvent(event);
-        }, selector);
-    };
-
-    /**
-     * Click an element using a JavaScript based event.
-     *
-     * @param {String} selector
-     */
-    click(selector) {
-        debug('.click() on ' + selector);
-        return this.evaluate_now(function (selector) {
-            document.activeElement.blur();
-            var element = document.querySelector(selector);
-            var event = document.createEvent('MouseEvent');
-            event.initEvent('click', true, true);
-            element.dispatchEvent(event);
-        }, selector);
-    };
-
-    /**
-     * Click an element and wait until the next load operation completes.
-     * Use this function when you expect a click to perform a navigation action, usually on anchor elements, but elsewhere too.
-     *
-     * @param {String} selector
-     */
-    clickAndWaitUntilFinishLoad(selector) {
-        debug('.clickAndWaitUntilFinishLoad() on ' + selector);
-
-        let child = this.child;
-        let waitUntilFinishLoadPromise = this._invokeRunnerOperation("waitUntilFinishLoad");
-
-        let clickPromise = this.evaluate_now(function (selector) {
-            document.activeElement.blur();
-            var element = document.querySelector(selector);
-            var event = document.createEvent('MouseEvent');
-            event.initEvent('click', true, true);
-            element.dispatchEvent(event);
-        }, selector);
-
-        return Promise.all([clickPromise, waitUntilFinishLoadPromise]);
-    };
-
-    /**
-     * Click an element using electron's sendInputEvent command.
-     *
-     * @param {String} selector
-     */
-    emulateClick(y, x) {
-        debug('.emulateClick() on ' + y);
-
-        //click the selector at y
-        if (_.isString(y) && !x) {
-            let self = this;
-            return co(function* () {
-                let clientRects = yield self.getClientRects(y);
-                let rect = clientRects[0];
-
-                let res = {
-                    x: Math.floor(rect.left + (rect.width / 2)),
-                    y: Math.floor(rect.top + (rect.height / 2))
-                };
-                debug('.emulateClick() found element at ' + res.x + ", " + res.y);
-                return self._invokeRunnerOperation("emulateClick", res);
-            });
-        }
-        //just pass the full object
-        else if (_.isObject(y) && !x) {
-            return this._invokeRunnerOperation("emulateClick", y);
-        }
-        //click x, y.
-        else {
-            return this._invokeRunnerOperation("emulateClick", { x: x, y: y });
-        }
-    }
-
-    /**
-     * Click an element using electron's sendInputEvent command.
-     *
-     * @param {String} selector
-     */
-    emulateKeystrokes(selector, text, opts) {
-        if (!text) {
-            text = selector;
-            selector = null;
-        }
-
-        opts = _.defaults(opts, {
-            initialFocusDelay: 750,
-            finalKeystrokeDelay: 500
-        });
-
-        debug('.emulateKeystrokes() on ' + selector);
-
-        let self = this;
-        return co(function* () {
-            if (selector) {
-                yield self.emulateClick(selector)
-                yield delay(opts.initialFocusDelay);
-            }
-
-            let keyCodes = [];
-            if (_.isArray(text))
-                keyCodes = text;
-            else {
-                for (let char of Array.from(text)) {
-                    let acc = {
-                        keyCode: char,
-                        modifiers: [],
-                        isChar: true
-                    };
-
-                    if (char == char.toUpperCase())
-                        acc.modifiers.push("Shift");
-
-                    keyCodes.push(acc);
-                }
-            }
-
-            return self._invokeRunnerOperation("emulateKeystrokes", { keyCodes: keyCodes, finalKeystrokeDelay: opts.finalKeystrokeDelay });
-        });
-    }
-
-    /**
      * end
      */
     end() {
@@ -416,25 +273,11 @@ class Nightmare {
     };
 
     /**
-     * Returns a promise which invokes the specified action which expects to perform a navigation action.
-     */
-    expectNavigation(fn, timeout) {
-        if (!timeout)
-            timeout = this._options.waitTimeout;
-        
-        let waitPromise = Promise.all([this.waitUntilFinishLoad(), fn.apply(this)]);
-
-        let timeoutPromise = new Promise(function (resolve, reject) {
-            setTimeout(reject, timeout, ".expectNavigation() timed out after " + timeout);
-        });
-        return Promise.race([waitPromise, timeoutPromise]);
-    };
-
-    /**
      * Returns a promise which will be resolved once the specified JavaScript has been evaluated.
      */
     evaluate_now(fn) {
         let args = Array.from(arguments).slice(1);
+        debug('.evaluate_now() fn on the page');
         let js = "(" + template.evaluate + "(" + (String(fn)) + ", " + (JSON.stringify(args)) + "))";
 
         var id = shortid.generate();
@@ -445,114 +288,20 @@ class Nightmare {
         return this._invokeRunnerOperation(operationParams, js, false, id);
     }
 
-    /**
-     * Evaluate a function on the page.
-     *
-     * @param {Function} fn
-     * @param {...} args
-     */
-    evaluate(fn/**, arg1, arg2...**/) {
-        let args = Array.from(arguments);
-        if (!_.isFunction(fn)) {
-            return done(new Error('.evaluate() fn should be a function'));
-        }
-        debug('.evaluate() fn on the page');
-        return this.evaluate_now.apply(this, args);
-    };
-
     /*
-     * Evaluate an asynchronous function on the page and wait for the promise/generator/thenable/callback to complete.
+     * Returns a promise which will be resolved when an asynchronous function on the page is evaluated and waits for the promise/generator/thenable/callback to complete.
      */
-    evaluateAsync(fn) {
+    evaluate_async(fn) {
         let args = Array.from(arguments).slice(1);
-        debug('.evaluateAsync() fn on the page');
+        debug('.evaluate_async() fn on the page');
         let js = "(" + template.evaluateAsync + "(" + (String(fn)) + ", " + (JSON.stringify(args)) + "))";
 
-        let child = this.child;
-        let p = new Promise(function (resolve, reject) {
-            child.once('javascript', function (result) {
-                if (result && !_.isUndefined(result.err)) {
-                    reject(result.err);
-                    return;
-                }
-
-                if (result && !_.isUndefined(result.result)) {
-                    resolve(result.result);
-                    return;
-                }
-
-                resolve(undefined);
-            });
-        });
-        this.child.emit('javascript', js, true);
-        return p;
-    }
-
-    /**
-     * Determine if a selector exists on a page.
-     *
-     * @param {String} selector
-     */
-    exists(selector) {
-        debug('.exists() for ' + selector);
-        return this.evaluate_now(function (selector) {
-            return (!!document.querySelector(selector));
-        }, selector);
-    };
-
-    /**
-     * Go forward to previous url.
-     */
-    forward() {
-        debug('.goForward()');
-
-        return this._invokeRunnerOperation("goForward");
-    };
-
-    /**
-     * Instructs the browser to go to a specific url and wait until loading completes.
-     * If the browser is currently at the specified URL, no action is taken.
-     */
-    goto(url, headers) {
-        debug('goto() starting navigation to %s', url);
-
-        headers = headers || {};
-        for (let key in this._headers) {
-            headers[key] = headers[key] || this._headers[key];
+        var id = shortid.generate();
+        var operationParams = {
+            runnerRequestMessageName: "javascript",
+            runnerResponseMessageName: "javascript " + id
         }
-
-        return this._invokeRunnerOperation("goto", url, headers);
-    };
-
-    getClientRects(selector) {
-
-        return this.evaluate_now(function (selector) {
-            'use strict';
-            var element = document.querySelector(selector);
-
-            if (!element)
-                throw "An element could not be located with the specified selector: " + selector;
-
-            var clientRects = element.getClientRects();
-
-            if (!clientRects)
-                throw "Client rects could not be retrieved."
-
-            var result = [];
-
-            for (let rect of Array.from(clientRects)) {
-                result.push({
-                    bottom: rect.bottom,
-                    height: rect.height,
-                    left: rect.left,
-                    right: rect.right,
-                    top: rect.top,
-                    width: rect.width
-                });
-            }
-
-            return result;
-        }, selector);
+        return this._invokeRunnerOperation(operationParams, js, true, id);
     }
 
     /**
@@ -566,14 +315,6 @@ class Nightmare {
         }
 
         return this._noop();
-    };
-
-    /*
-     * Save the contents of the current page as html
-     */
-    html(path, saveType) {
-        debug('.html() starting');
-        return this._invokeRunnerOperation("html", path, saveType);
     };
 
     /*
@@ -719,76 +460,12 @@ class Nightmare {
         }
         else if (type === 'css') {
             let css = fs.readFileSync(file, { encoding: 'utf-8' });
-            return this._invokeRunnerOperation("css", css);
+            return this._invokeRunnerOperation("insertCSS", css);
         }
         else {
             debug('unsupported file type in .inject()');
             throw new Error("unsupported file type in .inject(): " + type);
         }
-    };
-
-    /**
-     * Insert text
-     *
-     * @param {String} selector
-     * @param {String} text
-     */
-    insert(selector, text) {
-        debug('.insert() %s into %s', text, selector);
-
-        let self = this;
-        return co(function* () {
-            if (!text) {
-                return self.evaluate_now(function (selector) {
-                    document.querySelector(selector).focus();
-                    document.querySelector(selector).value = '';
-                }, selector);
-            } else {
-                try {
-                    yield self.evaluate_now(function (selector) {
-                        document.querySelector(selector).focus();
-                    }, selector);
-                }
-                catch (ex) {
-                    throw ex;
-                }
-
-                return self._invokeRunnerOperation("insert", text);
-            }
-        });
-    }
-
-
-    /**
-     * Mousedown on an element.
-     *
-     * @param {String} selector
-     */
-    mousedown(selector) {
-        debug('.mousedown() on ' + selector);
-        return this.evaluate_now(function (selector) {
-            var element = document.querySelector(selector);
-            var event = document.createEvent('MouseEvent');
-            event.initEvent('mousedown', true, true);
-            element.dispatchEvent(event);
-        }, selector);
-    };
-
-
-    /**
-     * Hover over an element.
-     *
-     * @param {String} selector
-     * @param {Function} done
-     */
-    mouseover(selector) {
-        debug('.mouseover() on ' + selector);
-        return this.evaluate_now(function (selector) {
-            var element = document.querySelector(selector);
-            var event = document.createEvent('MouseEvent');
-            event.initMouseEvent('mouseover', true, true);
-            element.dispatchEvent(event);
-        }, selector);
     };
 
     /**
@@ -803,217 +480,6 @@ class Nightmare {
     };
 
     /**
-     * Take a pdf.
-     *
-     * @param {String} path
-     * @param {Object} options
-     */
-    pdf(path, options) {
-        debug('.pdf()');
-        if (!options && _.isObject(path)) {
-            options = path;
-            path = undefined;
-        }
-
-        return this._invokeRunnerOperation("pdf", path, options)
-            .then(function (pdf) {
-                let buf = new Buffer(pdf.data);
-                debug('.pdf() captured with length %s', buf.length);
-                if (!path)
-                    return buf;
-
-                return fs.writeFileSync(path, buf);
-            });
-    };
-
-    /**
-     * Refresh the current page.
-     */
-    refresh() {
-        debug('.refresh()');
-        return this.evaluate_now(function () {
-            window.location.reload();
-        });
-    };
-
-    /**
-     * instructs the browser to reload the page.
-     */
-    reload() {
-        debug('.reload()');
-        return this._invokeRunnerOperation("reload");
-    }
-
-    /**
-     * Take a screenshot.
-     *
-     * @param {String} path
-     * @param {Object} clip
-     */
-    screenshot(path, clip) {
-        debug('.screenshot()');
-
-        if (!clip && _.isObject(path)) {
-            clip = path;
-            path = undefined;
-        }
-
-        return this._invokeRunnerOperation("screenshot", clip)
-            .then(function (img) {
-                let buf = new Buffer(img.data);
-                debug('.screenshot() captured with length %s', buf.length);
-                if (!path)
-                    return buf;
-
-                return fs.writeFileSync(path, buf);
-            });
-    };
-
-    /**
-     * Set the scroll position.
-     *
-     * @param {Number} x
-     * @param {Number} y
-     */
-    scrollTo(y, x) {
-        debug('.scrollTo()');
-
-        if (!x && _.isString(y)) {
-            return this.evaluate_now(function (selector) {
-                var element = document.querySelector(selector);
-                if (element) {
-                    var rect = element.getBoundingClientRect();
-                    window.scrollTo(Math.round(rect.left), Math.round(rect.top));
-                }
-                else
-                    throw 'invalid selector "' + selector + '"';
-            }, y);
-        }
-        else if (_.isNumber(x) && _.isNumber(x)) {
-            return this.evaluate_now(function (y, x) {
-                window.scrollTo(x, y);
-            }, y, x);
-        }
-    };
-
-    /**
-     * Choose an option from a select dropdown
-     *
-     * @param {String} selector
-     * @param {String} option value
-     */
-    select(selector, option) {
-        debug('.select() ' + selector);
-        return this.evaluate_now(function (selector, option) {
-            var element = document.querySelector(selector);
-            var event = document.createEvent('HTMLEvents');
-            element.value = option;
-            event.initEvent('change', true, true);
-            element.dispatchEvent(event);
-        }, selector, option);
-    };
-
-    /**
-     * Set the state of audio in the browser process.
-     *
-     * @param {bool} value that indicates if audio should be muted.
-     */
-
-    setAudioMuted(isMuted) {
-        debug('.setAudioMuted() to ' + isMuted);
-
-        return this._invokeRunnerOperation("audio", isMuted);
-    };
-
-    /**
-     * If prompted for login, supplies the specified credentials.
-     */
-    setAuthenticationCredentials(username, password) {
-        debug(".authentication()");
-
-        return this._invokeRunnerOperation("setAuthenticationCredentials", username, password);
-    };
-
-    /**
-     * instructs the browser to stop page loading.
-     */
-    stop() {
-        debug('.stop()');
-
-        return this._invokeRunnerOperation("stop");
-    }
-
-    /**
-     * Get the title of the page.
-     */
-    title() {
-        debug('.title() getting it');
-
-        return this._invokeRunnerOperation("title");
-    };
-
-    /**
-     * Type into an element.
-     *
-     * @param {String} selector
-     * @param {String} text
-     */
-    type() {
-        let selector = arguments[0], text;
-        if (arguments.length == 2) {
-            text = arguments[1];
-        }
-
-        debug('.type() %s into %s', text, selector);
-        let child = this.child;
-        let self = this;
-        return co(function* () {
-            if (!text) {
-                return self.evaluate_now(function (selector) {
-                    document.querySelector(selector).focus();
-                    document.querySelector(selector).value = '';
-                }, selector);
-            } else {
-                try {
-                    yield self.evaluate_now(function (selector) {
-                        document.querySelector(selector).focus();
-                    }, selector);
-                }
-                catch (ex) {
-                    throw ex;
-                }
-
-                return self._invokeRunnerOperation("type", text);
-            }
-        });
-    };
-
-    /*
-     * Uncheck a checkbox, fire change event
-     *
-     * @param {String} selector
-     */
-    uncheck(selector) {
-        debug('.uncheck() ' + selector);
-        return this.evaluate_now(function (selector) {
-            var element = document.querySelector(selector);
-            var event = document.createEvent('HTMLEvents');
-            element.checked = null;
-            event.initEvent('change', true, true);
-            element.dispatchEvent(event);
-        }, selector);
-    };
-
-    /**
-     * Get the url of the page.
-     */
-    url() {
-        debug('.url() getting it');
-
-        return this._invokeRunnerOperation("url");
-    };
-
-    /**
      * Provides an easy way to call fn.apply(nightmare)
      * @param {Function} 
      */
@@ -1022,115 +488,6 @@ class Nightmare {
         let fnPromise = co.wrap(fn);
         return fnPromise.apply(this);
     };
-
-    /**
-     * Set the useragent.
-     *
-     * @param {String} useragent
-     */
-    useragent(useragent) {
-        debug('.useragent() to ' + useragent);
-
-        return this._invokeRunnerOperation("useragent", useragent);
-    };
-
-    /**
-     * Set the viewport.
-     *
-     * @param {Number} width
-     * @param {Number} height
-     */
-    viewport(width, height) {
-        debug('.viewport()');
-
-        return this._invokeRunnerOperation("size", width, height);
-    };
-
-    /**
-     * Determine if a selector is visible on a page.
-     *
-     * @param {String} selector
-     */
-    visible(selector) {
-        debug('.visible() for ' + selector);
-        return this.evaluate_now(function (selector) {
-            var elem = document.querySelector(selector);
-            if (elem)
-                return (elem.offsetWidth > 0 && elem.offsetHeight > 0);
-            else
-                return false;
-        }, selector);
-    };
-
-    /**
-     * Wait
-     */
-    wait() {
-        let args = Array.from(arguments);
-        if (args.length < 1) {
-            debug('Not enough arguments for .wait()');
-            return;
-        }
-
-        let self = this;
-
-        let timeout = new Promise(function (resolve, reject) {
-            setTimeout(reject, self._options.waitTimeout, ".wait() timed out after " + self._options.waitTimeout);
-        });
-
-        let arg = args[0];
-        if (_.isNumber(arg)) {
-            debug('.wait() for ' + arg + 'ms');
-            return Promise.race([delay(arg), timeout]);
-        }
-        else if (_.isString(arg)) {
-            debug('.wait() for ' + arg + ' element');
-            let fn = function (selector) {
-                var element = document.querySelector(selector);
-                return (element ? true : false);
-            }
-            return this.waitUntilTrue(fn, arg);
-        }
-        else if (_.isFunction(arg)) {
-            debug('.wait() for fn');
-            return this.waitUntilTrue.apply(args);
-        }
-
-        throw new Error("wait() Unsupported Arguments.");
-    };
-
-    /**
-     * Returns a promise that resolves when the current web browser finishes loading all resources.
-     */
-    waitUntilFinishLoad() {
-        debug('waitUntilFinishLoad() starting');
-
-        return this._invokeRunnerOperation("waitUntilFinishLoad");
-    };
-
-    /**
-     * Wait until evaluated function returns true or timeout
-     *
-     * @param {Function} fn
-     * @param {...} args
-     */
-    waitUntilTrue(fn/**, arg1, arg2...**/) {
-        let args = Array.from(arguments);
-        let self = this;
-
-        let timeout = new Promise(function (resolve, reject) {
-            setTimeout(reject, self._options.waitTimeout, task.name + " timed out after " + self._options.waitTimeout);
-        });
-
-        let check = function* () {
-            let testResult = false;
-            do {
-                testResult = yield self.evaluate_now.apply(self, args);
-            } while (!testResult)
-        };
-
-        return Promise.race([check, timeout]);
-    }
 
     /*
      * Class properties
