+++ lib/runner.js
@@ -80,6 +80,7 @@ app.on('ready', function () {
                 nodeIntegration: false
             }
         });
+
         /**
          * Create a new Browser Window
          */
@@ -141,166 +142,6 @@ app.on('ready', function () {
     });
 
     /**
-     * Parent actions
-     */
-
-    /**
-     * goto
-     */
-
-    parent.on('goto', function (url, headers) {
-        var extraHeaders = '';
-        for (var key in headers) {
-            extraHeaders += key + ': ' + headers[key] + '\n';
-        }
-
-        if (win.webContents.getURL() == url) {
-            parent.emit('goto', {
-                result: url
-            });
-        } else {
-            var resolveGoto = function (message) {
-                win.webContents.removeListener("did-fail-load", rejectGoto);
-                parent.emit('goto', {
-                    result: win.webContents.getURL()
-                });
-            };
-            var rejectGoto = function (message) {
-                win.webContents.removeListener("did-finish-load", resolveGoto);
-                parent.emit('goto', {
-                    error: message
-                });
-            };
-
-            win.webContents.once('did-fail-load', rejectGoto);
-            win.webContents.once('did-finish-load', resolveGoto);
-
-            win.webContents.loadURL(url, {
-                extraHeaders: extraHeaders
-            });
-        }
-    });
-
-    /**
-     * back
-     **/
-    parent.on('goBack', function () {
-        if (!win.webContents.canGoBack()) {
-            parent.emit('goBack', {
-                error: true
-            });
-        } else {
-            win.webContents.once('did-finish-load', function () {
-                parent.emit('goBack', {
-                    result: win.webContents.getURL()
-                });
-            });
-            win.webContents.goBack();
-        }
-    });
-
-    /**
-     * forward
-     **/
-    parent.on('goForward', function () {
-        if (!win.webContents.canGoForward()) {
-            parent.emit('goForward', true);
-        } else {
-            win.webContents.once('did-finish-load', function () {
-                parent.emit('goForward', {
-                    result: win.webContents.getURL()
-                });
-            });
-            win.webContents.goForward();
-        }
-    });
-
-    /**
-     * stop
-     */
-    parent.on('stop', function () {
-        win.webContents.stop();
-        parent.emit('stop');
-    });
-
-    /**
-     * reload
-     */
-    parent.on('reload', function () {
-        win.webContents.reload();
-        parent.emit('reload');
-    });
-
-    /**
-     * Retrieves the specified element from clickOpts.selector and clicks it using webContents.sendInputEvent.
-     */
-    parent.on('emulateClick', function (clickOpts) {
-        clickOpts = _.defaults(clickOpts, {
-            button: "left",
-            clickCount: 1,
-            clickDelay: 25
-        });
-
-        var x = Math.round(clickOpts.x);
-        var y = Math.round(clickOpts.y);
-        parent.emit("log", "clicking " + x + ", " + y + ".");
-        win.webContents.sendInputEvent({ type: 'mouseDown', x: x, y: y, button: clickOpts.button, clickCount: clickOpts.clickCount });
-        setTimeout(function () {
-            win.webContents.sendInputEvent({ type: 'mouseUp', x: x, y: y, button: clickOpts.button, clickCount: clickOpts.clickCount });
-            parent.emit("emulateClick", {
-                result: { x: x, y: y }
-            });
-        }, clickOpts.clickDelay);
-        
-    });
-
-    parent.on('emulateKeystrokes', function (keystrokeOpts) {
-        keystrokeOpts = _.defaults(keystrokeOpts, {
-            keyCodes: [],
-            keystrokeDelay: 87,
-            finalKeystrokeDelay: 500
-        });
-
-        var q = async.queue(function (task, callback) {
-            parent.emit("log", task.keyCode);
-            win.webContents.sendInputEvent({
-                type: 'keyDown',
-                keyCode: task.keyCode,
-                modifier: task.modifier
-            });
-
-            if (task.isChar) {
-                win.webContents.sendInputEvent({
-                    type: 'char',
-                    keyCode: task.keyCode,
-                    modifier: task.modifier
-                });
-            }
-            
-            setTimeout(function () {
-                win.webContents.sendInputEvent({
-                    type: 'keyUp',
-                    keyCode: task.keyCode,
-                    modifiers: task.modifiers
-                });
-                callback();
-            }, keystrokeOpts.keystrokeDelay);
-            
-        }, 1);
-
-        q.drain = function () {
-            //this is to allow the final keyup to be fired.
-            setTimeout(function () {
-                parent.emit("emulateKeystrokes");
-            }, keystrokeOpts.finalKeystrokeDelay);
-        }
-
-        for (var keyCode of keystrokeOpts.keyCodes) {
-            q.push(keyCode)
-        }
-    });
-
-    /**
      * javascript
      */
     parent.on('javascript', function (code, isAsync, responseId) {
@@ -334,223 +175,11 @@ app.on('ready', function () {
     });
 
     /**
-     * css
+     * insert css
      */
-
-    parent.on('css', function (css) {
+    parent.on('insertCSS', function (css) {
         win.webContents.insertCSS(css);
-        parent.emit("css");
-    });
-
-    /**
-     * size
-     */
-
-    parent.on('size', function (width, height) {
-        win.setSize(width, height);
-        parent.emit("size");
-    });
-
-    /**
-     * title
-     */
-    parent.on('title', function () {
-        parent.emit("title", {
-            result: win.webContents.getTitle()
-        });
-    });
-
-    /**
-     * url
-     */
-    parent.on('url', function () {
-        parent.emit("url", {
-            result: win.webContents.getURL()
-        });
-    });
-
-    /**
-     * audio
-     */
-
-    parent.on('audio', function (audio) {
-        win.webContents.setAudioMuted(audio);
-        parent.emit('audio', {
-            result: win.webContents.isAudioMuted()
-        });
-    });
-
-    /**
-     * useragent
-     */
-    parent.on('useragent', function (useragent) {
-        win.webContents.setUserAgent(useragent);
-        parent.emit('useragent');
-    });
-
-    /**
-     * type
-     */
-
-    parent.on('type', function (value) {
-        var chars = String(value).split('')
-
-        function type() {
-            var ch = chars.shift()
-            if (ch === undefined) {
-                parent.emit('type');
-                return;
-            }
-
-            // keydown
-            win.webContents.sendInputEvent({
-                type: 'keyDown',
-                keyCode: ch
-            });
-
-            // keypress
-            win.webContents.sendInputEvent({
-                type: 'char',
-                keyCode: ch
-            });
-
-            // keyup
-            win.webContents.sendInputEvent({
-                type: 'keyUp',
-                keyCode: ch
-            });
-
-            // HACK to prevent async keyboard events from
-            // being played out of order. The timeout is
-            // somewhat arbitrary. I want to achieve a
-            // nice balance between speed and correctness
-            // if you find that this value it too low,
-            // please open an issue.
-            setTimeout(type, 100);
-        }
-
-        // start
-        type();
-    })
-
-    /**
-     * Authentication
-     */
-    parent.on('setAuthenticationCredentials', function (username, password) {
-        win.webContents.on('login', function (webContents, request, authInfo, callback) {
-            callback(username, password);
-        });
-        parent.emit('setAuthenticationCredentials');
-    });
-
-    /**
-     * Insert
-     */
-
-    parent.on('insert', function (value) {
-        win.webContents.insertText(String(value))
-        parent.emit('insert')
-    })
-
-    /**
-     * screenshot
-     */
-
-    parent.on('screenshot', function (clip) {
-        // https://gist.github.com/twolfson/0d374d9d7f26eefe7d38
-        var args = [function handleCapture(img) {
-            parent.emit('screenshot', {
-                result: img.toPng()
-            });
-        }];
-        if (clip) args.unshift(clip);
-        frameManager.requestFrame(function () {
-            win.capturePage.apply(win, args);
-        });
-    });
-
-   /**
-    * html
-    */
-    parent.on('html', function (path, saveType) {
-        // https://github.com/atom/electron/blob/master/docs/api/web-contents.md#webcontentssavepagefullpath-savetype-callback
-        saveType = saveType || 'HTMLComplete'
-        win.webContents.savePage(path, saveType, function (err) {
-            parent.emit('html', {
-                err: err
-            });
-        });
-    });
-
-    /**
-     * pdf
-     */
-
-    parent.on('pdf', function (path, options) {
-        // https://github.com/fraserxu/electron-pdf/blob/master/index.js#L98
-        options = _.defaults(options, {
-            marginType: 0,
-            printBackground: true,
-            printSelectionOnly: false,
-            landscape: false
-        });
-
-        win.webContents.printToPDF(options, function (err, data) {
-            if (err) return parent.emit('pdf', arguments);
-            parent.emit('pdf', {
-                result: data
-            });
-        });
-    });
-
-    /**
-     * Wait until Finish Load
-     */
-    parent.on("waitUntilFinishLoad", function () {
-
-        var start;
-        var init = new Promise(function (resolve, reject) {
-            start = resolve;
-        });
-
-        if (!win.webContents.isLoading()) {
-            init.then(function () {
-                return new Promise(function (resolve, reject) {
-                    win.webContents.once('did-start-loading', function () {
-                        resolve();
-                    });
-                });
-            });
-        }
-
-        init.then(function () {
-            return new Promise(function (resolve, reject) {
-
-                let resolveGoto = function (message) {
-                    win.webContents.removeListener("did-fail-load", rejectGoto);
-                    var url = win.webContents.getURL();
-                    resolve(url);
-                };
-
-                let rejectGoto = function (message) {
-                    win.webContents.removeListener("did-finish-load", resolveGoto);
-                    parent.emit('goto', message);
-                };
-
-                win.webContents.once('did-fail-load', rejectGoto);
-                win.webContents.once('did-finish-load', resolveGoto);
-            });
-        }).then(function (url) {
-            parent.emit('waitUntilFinishLoad', {
-                result: url
-            });
-        }, function (message) {
-            parent.emit('waitUntilFinishLoad', {
-                error: message
-            });
-        });
-
-        start();
+        parent.emit("insertCSS");
     });
 
     /**
@@ -563,7 +192,7 @@ app.on('ready', function () {
                 parent: parent
             });
 
-        fn(name, options, parent, win, renderer);
+        fn(name, options, parent, win, renderer, frameManager);
         parent.emit('electronAction ' + id);
     });
