+++ package.json
@@ -2,7 +2,7 @@
   "name": "nightmare",
   "version": "3.0.3",
   "license": "MIT",
-  "main": "lib/app.js",
+  "main": "app.js",
   "scripts": {
     "test": "make test",
     "coverage": "./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- "
